Create a "view" feature.
