+++ rs-lib/src/lib.rs
@@ -1,25 +1,23 @@
 mod comments;
-mod custom;
 mod pos;
 mod text_info;
 mod tokens;
 mod types;
 
-#[allow(clippy::all)]
-#[rustfmt::skip]
-mod generated;
-
 #[cfg(test)]
-mod test_helpers;
+pub(crate) mod test_helpers;
 
-pub use comments::CommentsIterator;
-pub use custom::*;
-pub use generated::*;
-pub use types::*;
+#[cfg(feature = "view")]
+mod view;
+#[cfg(feature = "view")]
+pub use view::*;
 
+pub use comments::*;
 pub use pos::*;
 pub use text_info::*;
 pub use text_lines;
+pub use tokens::*;
+pub use types::*;
 
 /// A 0-indexed line and column type.
 pub type LineAndColumnIndex = text_lines::LineAndColumnIndex;
